+++ lab8/src/response.h
@@ -6,12 +6,12 @@
 const unsigned int PAYLOAD=800;
 
 typedef struct __attribute__((packed))response{
-    uint16_t seq;
+    uint32_t seq;
     uint32_t checksum;
     uint8_t payload[PAYLOAD];
 	uint8_t flag;
     response();
-    response(uint16_t,uint8_t*,unsigned);
+    response(uint32_t,uint8_t*,unsigned);
     uint32_t genHash();
     bool check();
 }response;
